Add files via upload
Apollo Hospitals was established in 1983, renowned as the architect of modern healthcare in India. As the nation's first corporate hospital, Apollo Hospitals is acclaimed for pioneering the private healthcare revolution in the country. As a data scientist working at Apollo 24/7, the ultimate goal is to tease out meaningful and actionable insights from Patient-level collected data. You can help Apollo hospitals to be more efficient, to influence diagnostic and treatment processes, to map the spread of a pandemic.

One of the best examples of data scientists making a meaningful difference at a global level is in the response to the COVID-19 pandemic, where they have improved information collection, provided ongoing and accurate estimates of infection spread and health system demand, and assessed the effectiveness of government policies.
How can you help here?

The company wants to know:
• Which variables are significant in predicting the reason for hospitalization for different regions
• How well some variables like viral load, smoking, Severity Level describe the hospitalization charges

Dataset: https://d2beiqkhq929f0.cloudfront.net/public_assets/assets/000/001/681/original/scaler_apollo_hospitals.csv

Column Profiling
Age: This is an integer indicating the age of the primary beneficiary (excluding those above 64 years, since they are generally covered by the government).
Sex: This is the policy holder's gender, either male or female
Viral Load: Viral load refers to the amount of virus in an infected person's blood
Severity Level: This is an integer indicating how severe the patient is
Smoker: This is yes or no depending on whether the insured regularly smokes tobacco.
Region: This is the beneficiary's place of residence in Delhi, divided into four geographic regions - northeast, southeast, southwest, or northwest
Hospitalization charges: Individual medical costs billed to health insurance

Concept Used:
Graphical and Non-Graphical Analysis
2-sample t-test: testing for difference across populations
ANOVA
Chi-square